fix(security): redact encoded URL query tokens
