+++ src/lib/security/redact.test.ts
@@ -3,159 +3,7 @@
 
 import { describe, expect, it } from "vitest";
 
-import {
-  redact,
-  redactForLog,
-  redactFull,
-  redactLogSequence,
-  redactSensitiveText,
-  redactUrl,
-} from "./redact.js";
-
-function makeJwtFixture(): string {
-  return ["eyJhbGciOiJIUzI1NiJ9", "eyJzdWIiOiIxMjM0NTY3ODkwIn0", "signatureABCDEFGHI"].join(".");
-}
-
-function makeEmptyClaimsJwtFixture(): string {
-  return ["eyJhbGciOiJIUzI1NiJ9", "e30", "signatureABCDEFGHI"].join(".");
-}
-
-describe("URL redaction", () => {
-  it.each([
-    ["SOCKS", "socks5://socks-user:socks-password@proxy.example:1080"],
-    ["mixed-case FTP", "FtP://ftp-user:ftp-password@files.example/path"],
-    ["mixed-case HTTPS", "HTTPS://https-user:https-password@secure.example:8443"],
-  ])("redacts embedded credentials from %s URLs", (_label, value) => {
-    const result = redact(value);
-
-    expect(result).toContain("****:****@");
-    expect(result).not.toContain("-user");
-    expect(result).not.toContain("-password");
-  });
-
-  it("redacts a bracket-wrapped SOCKS URL without breaking its closing delimiter", () => {
-    const result = redact(
-      "proxy [socks5://bracket-user:bracket-password@proxy.example:1080] failed",
-    );
-
-    expect(result).toContain("socks5://****:****@proxy.example:1080]");
-    expect(result).not.toContain("bracket-user");
-    expect(result).not.toContain("bracket-password");
-  });
-
-  it("bounds malformed wrapper parsing before falling back to userinfo redaction", () => {
-    const wrappers = "]".repeat(4_096);
-    const result = redact(
-      `proxy [socks5://bounded-user:bounded-password@proxy.example:1080${wrappers}`,
-    );
-
-    expect(result).toContain("socks5://****:****@proxy.example:1080");
-    expect(result).not.toContain("bounded-user");
-    expect(result).not.toContain("bounded-password");
-  });
-
-  it("preserves a credentialed IPv6 host while redacting its userinfo", () => {
-    const result = redact("proxy https://ipv6-user:ipv6-password@[::1]:8443/path failed");
-
-    expect(result).toContain("https://****:****@[::1]:8443/path");
-    expect(result).not.toContain("ipv6-user");
-    expect(result).not.toContain("ipv6-password");
-  });
-
-  it.each([
-    [
-      "parentheses and comma",
-      "proxy (https://wrapped-user:wrapped-password@proxy.example/path), retry",
-      "(https://****:****@proxy.example/path), retry",
-    ],
-    [
-      "angle brackets and semicolon",
-      "proxy <ftp://wrapped-user:wrapped-password@files.example/path>; retry",
-      "<ftp://****:****@files.example/path>; retry",
-    ],
-    [
-      "a trailing sentence period",
-      "proxy socks5://wrapped-user:wrapped-password@proxy.example:1080. retry",
-      "socks5://****:****@proxy.example:1080. retry",
-    ],
-  ])("keeps %s outside the redacted URL token", (_label, value, expected) => {
-    const result = redact(value);
-
-    expect(result).toContain(expected);
-    expect(result).not.toContain("wrapped-user");
-    expect(result).not.toContain("wrapped-password");
-  });
-
-  it.each([
-    ["semicolon", "pa;ssword"],
-    ["comma", "pa,ssword"],
-    ["balanced parentheses", "pa(ss)word"],
-  ])("redacts credentials containing valid %s punctuation", (_label, password) => {
-    const result = redact(`proxy https://userinfo-user:${password}@proxy.example/path failed`);
-
-    expect(result).toContain("https://****:****@proxy.example/path");
-    expect(result).not.toContain("userinfo-user");
-    expect(result).not.toContain(password);
-  });
-
-  it("fully removes generic-scheme userinfo and sensitive query values", () => {
-    const result = redactUrl(
-      "FtP://ftp-user:ftp-password@files.example/path?token=secret-value#fragment",
-    );
-
-    expect(result).toBe("ftp://files.example/path?token=%3CREDACTED%3E");
-  });
-
-  it.each([
-    "api_key",
-    "apiKey",
-    "password",
-    "secret",
-    "authorization",
-    "credential",
-  ])("redacts opaque query values under credential key %s", (key) => {
-    const result = redactUrl(`https://endpoint.example/api?${key}=opaque-value`);
-
-    expect(result).not.toBeNull();
-    expect(new URL(result as string).searchParams.get(key)).toBe("<REDACTED>");
-  });
-
-  it("redacts encoded and repeated token-shaped query values under benign names", () => {
-    const secrets = {
-      anthropic: "sk-ant-abcdefghijklmnopqrstuvwxyz",
-      github: "ghp_abcdefghijklmnopqrstuvwxyz",
-      jwt: makeJwtFixture(),
-      jwtEmptyClaims: makeEmptyClaimsJwtFixture(),
-      nvidia: "nvapi-abcdefghijklmnopqrstuvwxyz",
-      openai: "sk-proj-abcdefghijklmnopqrstuvwxyz",
-      slack: "xoxb-1234567890-abcdefghij",
-    };
-    const encodedOpenAi = secrets.openai.replaceAll("-", "%2D");
-    const result = redactUrl(
-      `https://url-user:url-password@endpoint.example/api?value=${secrets.nvidia}&model=${encodedOpenAi}&provider=${secrets.anthropic}&session=${secrets.github}&channel=${secrets.slack}&assertion=${secrets.jwt}&compact=${secrets.jwtEmptyClaims}&repeat=safe&repeat=${secrets.github}&mixed=prefix:${secrets.nvidia}:suffix&token=opaque-value&keep=yes#session=${secrets.slack}`,
-    );
-
-    expect(result).not.toBeNull();
-    const parsed = new URL(result as string);
-    expect(parsed.username).toBe("");
-    expect(parsed.password).toBe("");
-    expect(parsed.hash).toBe("");
-    expect(parsed.searchParams.get("value")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("model")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("provider")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("session")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("channel")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("assertion")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("compact")).toBe("<REDACTED>");
-    expect(parsed.searchParams.getAll("repeat")).toEqual(["safe", "<REDACTED>"]);
-    expect(parsed.searchParams.get("mixed")).toBe("prefix:<REDACTED>:suffix");
-    expect(parsed.searchParams.get("token")).toBe("<REDACTED>");
-    expect(parsed.searchParams.get("keep")).toBe("yes");
-    for (const secret of Object.values(secrets)) {
-      expect(result).not.toContain(secret);
-    }
-  });
-});
+import { redactForLog, redactFull, redactLogSequence, redactSensitiveText } from "./redact.js";
 
 describe("redactForLog", () => {
   it("redacts pass aliases in structured keys and canonical text assignments", () => {
